schedule implementation

+++ src/app/shared/admin/department/department.component.ts
@@ -5,6 +5,7 @@ import * as alertify from 'alertifyjs';
 import { NgxPaginationModule } from 'ngx-pagination';
 
 import { CommonModule } from '@angular/common';
+import { ConfirmationService } from '../../confirmation/confirmation.service';
 
 @Component({
   selector: 'app-department',
@@ -23,7 +24,9 @@ export class DepartmentComponentimplements {
 
 
 
-  constructor(private formBuilder: FormBuilder, private departmentService: DepartmentService) { }
+  constructor(private formBuilder: FormBuilder, private departmentService: DepartmentService
+    ,private confirmationService:ConfirmationService
+  ) { }
   department: any[] = []
   ngOnInit(): void {
     this.departmentForm = this.formBuilder.group({
@@ -52,20 +55,28 @@ export class DepartmentComponentimplements {
     this.searchTerm = '';
     this.filterDepartments();
   }
-  deleteDepartment(id: string) {
-    this.departmentService.deleteDepartment(id).subscribe(
-      (response) => {
-        alertify.success('Successfully Deleted')
-        this.getDepartmentList()
-
-        console.log('Department deleted:', response);
-      },
-      error => {
-        console.error('Error deleting department:', error);
-        this.getDepartmentList()
-      }
-    );
+  async deleteDepartment(id: string) {
+    console.log('button isclicked');
+    const confirmed = await this.confirmationService.showConfirmationPopup();
+    debugger
+    if (confirmed) {
+      this.departmentService.deleteDepartment(id).subscribe(
+        (response) => {
+          this.getDepartmentList();
+          this.confirmationService.showSuccessMessage('Delete Successfully Done');
+          console.log('Department deleted:', response);
+        },
+        (error) => {
+          this.confirmationService.showErrorMessage('Sorry, Cannot be Deleted');
+          this.getDepartmentList();
+          console.error('Error deleting department:', error);
+        }
+      );
+    } else {
+      this.confirmationService.showErrorMessage('Delete operation cancelled');
+    }
   }
+  
   onSubmit(): void {
     if (this.departmentForm.valid) {
       const formData = this.departmentForm.value;
@@ -73,6 +84,7 @@ export class DepartmentComponentimplements {
         this.departmentService.updateDepartment(this.editingDepartmentId, formData).subscribe(
           (res) => {
             console.log('Department updated successfully:', res);
+            this.confirmationService.showSuccessMessage('Department updated successfully');
             this.editingDepartmentId = null;
             this.departmentForm.reset();
             this.getDepartmentList();
@@ -85,16 +97,22 @@ export class DepartmentComponentimplements {
         this.departmentService.postDepartment(formData).subscribe(
           (res) => {
             console.log('Department added successfully:', res);
+            this.confirmationService.showSuccessMessage('Department added successfully');
+
             this.departmentForm.reset();
             this.getDepartmentList();
           },
           (error) => {
             console.error('Error adding department:', error);
+            this.confirmationService.showErrorMessage('Error adding department');
           }
         );
       }
     } else {
       console.error('Form is invalid.');
+      this.confirmationService.showErrorMessage('Form is invalid');
+
+      
     }
   }
